Updated mongodb connection
Made it so that it connects to a free mongodb hosting. This is because to get various paypal features to work this app needs to be hosted on a live server. This requires that everything be hosted live including the mongodb. App is hosted on heroku, db is hosted on free mongodb Atlas servers.