Edited checkout.js, checkout route.
Order.create() should add an order object to the database. Please check this is working as intended @Nicole/@Josh. The code is in a typescript file and its just an interface so I wouldn't worry about it to much just try and find out which function we need to call. I think its Order.create() based on what I read.